+++ tests/ReplicationTest.php
@@ -4,7 +4,6 @@ namespace Phlib\DbHelperReplication;
 
 use Phlib\DbHelperReplication\Replication\StorageInterface;
 use Phlib\Db\AdapterInterface;
-use Phlib\DbHelperReplication\Replication\StorageMock;
 use phpmock\phpunit\PHPMock;
 use PHPUnit\Framework\TestCase;
 
@@ -42,56 +41,6 @@ class ReplicationTest extends TestCase
         parent::tearDown();
         $this->storage = null;
         $this->master  = null;
-    }
-
-    public function testCreateFromConfigSuccessfully()
-    {
-        $config = $this->getDefaultConfig();
-        $replication = Replication::createFromConfig($config);
-        static::assertInstanceOf(Replication::class, $replication);
-    }
-
-    /**
-     * @expectedException \Phlib\DbHelperReplication\Exception\InvalidArgumentException
-     */
-    public function testCreateFromConfigWithInvalidStorageClass()
-    {
-        $config = $this->getDefaultConfig();
-        $config['storage']['class'] = '\My\Unknown\Class';
-        Replication::createFromConfig($config);
-    }
-
-    /**
-     * @expectedException \Phlib\DbHelperReplication\Exception\InvalidArgumentException
-     */
-    public function testCreateFromConfigWithInvalidStorageMethod()
-    {
-        $config = $this->getDefaultConfig();
-        $config['storage']['class'] = \stdClass::class;
-        Replication::createFromConfig($config);
-    }
-
-    /**
-     * @return array
-     */
-    public function getDefaultConfig()
-    {
-        return [
-            'host'     => '10.0.0.1',
-            'username' => 'foo',
-            'password' => 'bar',
-            'slaves'   => [
-                [
-                    'host'     => '10.0.0.2',
-                    'username' => 'foo',
-                    'password' => 'bar'
-                ]
-            ],
-            'storage' => [
-                'class' => StorageMock::class,
-                'args'  => [[]]
-            ],
-        ];
     }
 
     /**