+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   directly, without `--daemonize`.
 ### Changed
 - Use `ext-memcached` in place of `ext-memcache`.
+- Move `Replication::createFromConfig()` static method to an instance method on
+  new class `ReplicationFactory`.
 ### Removed
 - Removed `replication:stats` CLI command which was not functional.
 - Removed support for PHP versions < v7.1 as they are no longer